refactor: separating interceptor to different file
